+++ hir/src/ops/array/add_dims.rs
@@ -44,7 +44,7 @@ impl Expansion for AddDims {
     ) -> InferenceResult {
         check_output_arity(&outputs, 1)?;
         s.equals(&outputs[0].datum_type, &inputs[0].datum_type)?;
-        s.equals(&outputs[0].rank, (&inputs[0].rank).bex() + self.axes.len() as i32)?;
+        s.equals(&outputs[0].rank, (&inputs[0].rank).bex() + self.axes.len() as i64)?;
         s.given(&inputs[0].shape, move |s, shape| {
             let output_shape = self.compute_shape(&shape);
             s.equals(&outputs[0].shape, output_shape)
